# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DentalAI SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dental/procedures:
    get:
      operationId: listProcedures
      tags: [procedures]
      summary: List dental procedures
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: List of procedures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DentalProcedure"
    post:
      operationId: createProcedure
      tags: [procedures]
      summary: Create a procedure
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcedureBody"
      responses:
        "201":
          description: Created procedure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DentalProcedure"
...
  /dental/appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: date
          in: query
          schema:
            type: string
            description: YYYY-MM-DD filter date
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithPatient"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Conflict - time slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
...
  /dental/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: temperature
          in: query
          schema:
            type: string
            enum: [cold, warm, hot]
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DentalLead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DentalLead"
...
  /dental/reports/appointments:
    get:
      operationId: getAppointmentsReport
      tags: [reports]
      summary: Appointments by period report
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        "200":
          description: Appointments report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsReport"
  /dental/reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Revenue report by period
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /dental/reports/procedures:
    get:
      operationId: getProceduresReport
      tags: [reports]
      summary: Top procedures report
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Procedures report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcedureReportItem"
  /dental/reports/peak-hours:
    get:
      operationId: getPeakHoursReport
      tags: [reports]
      summary: Peak hours report
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Peak hours data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeakHourItem"
  /dental/reports/leads:
    get:
      operationId: getLeadsReport
      tags: [reports]
      summary: Lead conversion report
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads conversion report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsReport"

  # ──────────────────────────── SETTINGS ────────────────────────────
...
components:
  parameters:
...
    TenantHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: integer
      description: Tenant (clinic) ID
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        cpf:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        profilePicUrl:
          type: string
          nullable: true
        firstVisit:
          type: string
          format: date-time
          nullable: true
        lastVisit:
          type: string
          format: date-time
          nullable: true
        totalSpent:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - phone
        - totalSpent
        - createdAt
        - updatedAt
...
    DentalProcedure:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        durationMinutes:
          type: integer
        active:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - price
        - durationMinutes
        - active
        - createdAt
        - updatedAt
    CreateProcedureBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        durationMinutes:
          type: integer
          default: 60
      required:
        - name
        - price
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        patientId:
          type: integer
          nullable: true
        leadId:
          type: integer
          nullable: true
        procedureId:
          type: integer
          nullable: true
        procedureName:
          type: string
          nullable: true
        status:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        price:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        confirmed:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - status
        - startsAt
        - endsAt
        - confirmed
        - createdAt
        - updatedAt
    AppointmentWithPatient:
      allOf:
        - $ref: "#/components/schemas/Appointment"
        - type: object
          properties:
            patient:
              $ref: "#/components/schemas/Patient"
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
          nullable: true
        leadId:
          type: integer
          nullable: true
        procedureId:
          type: integer
          nullable: true
        procedureName:
          type: string
          nullable: true
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        price:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - startsAt
        - endsAt
...
    ConflictError:
      type: object
      properties:
        error:
          type: string
        conflictingAppointment:
          $ref: "#/components/schemas/Appointment"
      required:
        - error

    # ─── LEAD ───
    DentalLead:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        temperature:
          type: string
          enum: [cold, warm, hot]
        source:
          type: string
          nullable: true
        interest:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        profilePicUrl:
          type: string
          nullable: true
        status:
          type: string
        convertedToPatientId:
          type: integer
          nullable: true
        convertedAt:
          type: string
          format: date-time
          nullable: true
        lastContactAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - phone
        - temperature
        - status
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        temperature:
          type: string
          enum: [cold, warm, hot]
          default: cold
        source:
          type: string
          nullable: true
        interest:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        professionalId:
          type: integer
          nullable: true
      required:
        - name
        - phone
...
    AppointmentsReport:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              count:
                type: integer
              completed:
                type: integer
              cancelled:
                type: integer
            required:
              - period
              - count
              - completed
              - cancelled
        total:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
      required:
        - data
        - total
        - completed
        - cancelled
    RevenueReport:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              revenue:
                type: number
            required:
              - period
              - revenue
        totalRevenue:
          type: number
        averagePerAppointment:
          type: number
      required:
        - data
        - totalRevenue
        - averagePerAppointment
    ProcedureReportItem:
      type: object
      properties:
        procedureName:
          type: string
        count:
          type: integer
        revenue:
          type: number
      required:
        - procedureName
        - count
        - revenue
    PeakHourItem:
      type: object
      properties:
        hour:
          type: integer
        count:
          type: integer
      required:
        - hour
        - count
    LeadsReport:
      type: object
      properties:
        total:
          type: integer
        converted:
          type: integer
        conversionRate:
          type: number
        byTemperature:
          type: object
          properties:
            cold:
              type: integer
            warm:
              type: integer
            hot:
              type: integer
          required:
            - cold
            - warm
            - hot
      required:
        - total
        - converted
        - conversionRate
        - byTemperature

    # ─── PROFESSIONALS ───
...